requirements.txt.txt
streamlit==1.28.1
requests==2.31.0
agno==0.1.16
firecrawl-py==0.0.16
pydantic==2.4.2
pandas==2.1.1